refactor(commands): extract speech-mode commands into speech.rs
